Feat: moved rate-limiter to own file
