+++ app.js
@@ -25,4 +25,4 @@ app.use(requestLogger);
 app.use(router);
 app.use(errorLogger);
 app.use(errors());
-app.use(error);
+// app.use(error);
