+++ routes/index.js
@@ -1,5 +1,4 @@
 const router = require('express').Router();
-const rateLimit = require('express-rate-limit');
 
 const { register, login } = require('../controllers/user');
 const NotFoundError = require('../errors/NotFoundError');
@@ -7,14 +6,7 @@ const auth = require('../middlewares/auth');
 const userRouter = require('./user');
 const movieRouter = require('./movie');
 const { registerValidation, loginValidation } = require('../middlewares/validation');
-
-//* * ограничитель запросов */
-const authLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 5, // Limit each IP to 100 requests per `window` (here, per 15 minutes)
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-  message: 'Слишком много запросов, пожалуйста попробуйте позже',
-});
+const authLimiter = require('../helpers/rate-limiter');
 
 //* * роуты не требующие авторизации */
 router.post('/signup', authLimiter, registerValidation, register);
